Move player logic to Player class
